+++ src/common/types.ts
@@ -5,7 +5,6 @@ export enum CandidateState {
 }
 
 export interface Candidate {
-    id: string;
     fullName: string;
     email: string;
     password: string;
